+++ train_model_methods/svc_train_classifier.py
